pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "credit-modelling"
version = "0.1.0"
description = "End-to-End MLOps Platform for Credit Risk Modelling"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "boto3>=1.30.0,<2.0",
    "mlflow>=2.10.0,<3.0",
    "openpyxl>=3.1.0,<4.0",
    "pandas>=2.1.0,<3.0",
    "plotly>=5.0.0,<6.0",
    "pyyaml>=6.0.0,<7.0",
    "ruff>=0.1.0",
    "scikit-learn>=1.3.0,<2.0",
    "streamlit>=1.30.0,<2.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["src"]
combine-as-imports = true
lines-after-imports = 2